+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/ClassRestController.java
@@ -3,10 +3,7 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.ClassDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.ClassController;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.ClassS;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/class")
@@ -22,6 +19,12 @@ public class ClassRestController {
     public ClassS createClass(@RequestBody ClassDTO classDTO){
             return classController.createClass(classDTO);
     }
+
+    @GetMapping("/{classId}")
+    public ClassS getClass(@PathVariable Long classId){
+        return classController.getClassById(classId);
+    }
 }
+