+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/ClassController.java
@@ -19,4 +19,10 @@ public class ClassController {
         ClassS classS = new ClassS(classDTO);
         return classService.saveClass(classS);
     }
+
+    public ClassS getClassById(Long classId) {
+        return classService.getClassById(classId);
+    }
+
+
 }